Use 64 bit precision

+++ serial/src/math.rs
@@ -1,5 +1,5 @@
 pub use cgmath::*;
 
-pub type Real = f32;
-pub type V2 = Vector2<f32>;
+pub type Real = f64;
+pub type V2 = Vector2<f64>;
 pub type V2i = Vector2<i32>;
